Refactor context algorithm implementation
