+++ clinlp/__init__.py
@@ -1,4 +1,4 @@
 from .language import Clinlp
 from .normalizer import Normalizer
-from .qualifier import ContextMatcher
+from .qualifier import ContextAlgorithm
 from .sentencizer import Sentencizer
