+++ clinlp/qualifier/__init__.py
@@ -1,11 +1,2 @@
-from spacy.tokens import Span
-
+from .context_algorithm import ContextAlgorithm, ContextRule, ContextRuleDirection
 from .qualifier import Qualifier, QualifierDetector
-from .rule_based import (
-    QUALIFIERS_ATTR,
-    ContextMatcher,
-    ContextRule,
-    ContextRuleDirection,
-)
-
-Span.set_extension(name=QUALIFIERS_ATTR, default=None)
