+++ clinlp/qualifier/context_algorithm.py
@@ -1,6 +1,3 @@
-""" Implements qualifier detection for entities. The Qualifier class is reusable. The other classes implement
-the context algorithm (https://doi.org/10.1016%2Fj.jbi.2009.05.002) """
-
 import importlib
 import itertools
 import json
@@ -16,17 +13,15 @@ from spacy.language import Language
 from spacy.matcher import Matcher, PhraseMatcher
 from spacy.tokens import Doc, Span
 
-from clinlp.qualifier.qualifier import Qualifier, QualifierDetector
+from clinlp.qualifier.qualifier import QUALIFIERS_ATTR, Qualifier, QualifierDetector
 
-QUALIFIERS_ATTR = "qualifiers"
+_DEFAULT_CONTEXT_RULES_PATH = str(importlib.resources.path("clinlp.resources", "psynlp_context_rules.json"))
 PHRASE_MATCHER_ATTR = "TEXT"
-
-DEFAULT_CONTEXT_RULES = "psynlp_context_rules.json"
 
 
 class ContextRuleDirection(Enum):
     """
-    Direction of a Context rule, as in the original algorithm.
+    Direction of a Context rule, as in the original Context Algorithm.
     """
 
     PRECEDING = 1
@@ -38,13 +33,13 @@ class ContextRuleDirection(Enum):
 @dataclass
 class ContextRule:
     """
-    A Context rule, as in the original algorithm.
+    A Context rule, as in the original Context Algorithm.
 
     Args:
-        pattern: The pattern to look for in text, either a string, or a spacy pattern (list).
-        qualifier: The qualifier to modify.
-        direction: The context rule direction.
-        max_scope: The maximum scope (number of tokens) of the trigger, within sentence boundaries
+        pattern: The pattern to look for in text. Either a string, or a spacy pattern (list).
+        qualifier: The qualifier to apply.
+        direction: The Context rule direction.
+        max_scope: The maximum scope (number of tokens) of the trigger, or None for using sentence boundaries.
     """
 
     pattern: Union[str, list[dict[str, str]]]
@@ -55,7 +50,7 @@ class ContextRule:
 
 class _MatchedContextPattern:
     """
-    A matched context pattern, that should be processed further.
+    A matched Context pattern, that should be processed further.
     """
 
     def __init__(self, rule: ContextRule, start: int, end: int, offset: int = 0):
@@ -64,7 +59,12 @@ class _MatchedContextPattern:
         self.end = end + offset
         self.scope = None
 
-    def set_initial_scope(self, sentence: Span):
+    def initialize_scope(self, sentence: Span):
+        """
+        Sets the scope this pattern ranges over, based on the sentence. This is either the window determined in
+        the `max_scope` of the rule, or the sentence boundaries if no `max_scope` is set.
+        """
+
         max_scope = self.rule.max_scope or len(sentence)
 
         if max_scope < 1:
@@ -77,84 +77,26 @@ class _MatchedContextPattern:
             self.scope = (max(self.start - max_scope, sentence.start), self.end)
 
 
-def _parse_qualifier(qualifier: str, qualifier_classes: dict[str, Qualifier]) -> Qualifier:
-    """
-    Parse a Qualifier from string.
-
-    Args:
-        qualifier: The qualifier (e.g. Negation.NEGATED).
-        qualifier_classes: A mapping of string to qualifier class.
-
-    Returns: A qualifier, as specified.
-    """
-
-    match_regexp = r"\w+\.\w+"
-
-    if not re.match(match_regexp, qualifier):
-        raise ValueError(
-            f"Cannot parse qualifier {qualifier}, please adhere to format "
-            f"{match_regexp} (e.g. NegationQualifier.NEGATED)"
-        )
-
-    qualifier_class, qualifier = qualifier.split(".")
-
-    return qualifier_classes[qualifier_class][qualifier]
-
-
-def _parse_direction(direction: str) -> ContextRuleDirection:
-    """
-    Parse a Context direction.
-
-    Args:
-        direction: The direction.
-
-    Returns: THe ContextRuleDirection.
-    """
-    return ContextRuleDirection[direction.upper()]
-
-
-def parse_rules(input_json: Optional[str] = None, data: Optional[dict] = None) -> list[ContextRule]:
-    if input_json and data:
-        raise ValueError(
-            "Please choose either input_json to load data from json, " "or provide data as dict, but not both."
-        )
-
-    if input_json:
-        with open(input_json, "rb") as file:
-            data = json.load(file)
-
-    qualifiers = {
-        qualifier["qualifier"]: Qualifier(qualifier["qualifier"], qualifier["levels"])
-        for qualifier in data["qualifiers"]
-    }
-
-    qualifier_rules = []
-
-    for rule in data["rules"]:
-        qualifier = _parse_qualifier(rule["qualifier"], qualifiers)
-        direction = _parse_direction(rule["direction"])
-        max_scope = rule.get("max_scope", None)
-
-        qualifier_rules += [ContextRule(pattern, qualifier, direction, max_scope) for pattern in rule["patterns"]]
-
-    return qualifier_rules
-
-
 @Language.factory(
-    name="clinlp_context_matcher",
-    default_config={"phrase_matcher_attr": PHRASE_MATCHER_ATTR, "rules": None},
+    name="clinlp_context_algorithm",
+    default_config={
+        "phrase_matcher_attr": PHRASE_MATCHER_ATTR,
+        "rules": _DEFAULT_CONTEXT_RULES_PATH,
+        "load_rules": True,
+    },
     requires=["doc.sents", "doc.ents"],
 )
-class ContextMatcher(QualifierDetector):
+class ContextAlgorithm(QualifierDetector):
     """
-    Implements a very simple version of the context algorithm.
+    Implements the Context algorithm (https://doi.org/10.1016%2Fj.jbi.2009.05.002) as a spaCy pipeline.
 
     Args:
         nlp: The Spacy language object to use
         name: The name of the component
         phrase_matcher_attr: The token attribute to match phrases on (e.g. TEXT, ORTH, NORM).
-        default_rules: A filename in clinlp.resources with a set of rules to load by default
-        rules: A list of ContextRule
+        rules: A dictionary of rules, or a path to a json containing the rules (see clinlp.resources dir for example).
+        load_rules: Whether to parse any rules. Set this to `False` to use ContextAlgorithm.add_rules to
+        add ContextRules manually.
     """
 
     def __init__(
@@ -162,8 +104,8 @@ class ContextMatcher(QualifierDetector):
         nlp: Language,
         name: str,
         phrase_matcher_attr: str = PHRASE_MATCHER_ATTR,
-        default_rules: Optional[str] = DEFAULT_CONTEXT_RULES,
-        rules: Optional[list[ContextRule]] = None,
+        load_rules=True,
+        rules: Optional[Union[str | dict]] = _DEFAULT_CONTEXT_RULES_PATH,
     ):
         self._nlp = nlp
         self.name = name
@@ -173,15 +115,71 @@ class ContextMatcher(QualifierDetector):
 
         self.rules = {}
 
-        if default_rules is not None:
-            self._load_default_rules(default_rules)
+        if load_rules:
+            if rules is None:
+                raise ValueError(
+                    "Did not provide rules. Set `parse_rules` to False if you " "want to add `ContextRule` manually."
+                )
 
-        if rules:
+            rules = self._parse_rules(rules)
             self.add_rules(rules)
 
-    def _load_default_rules(self, default_rules: str):
-        with importlib.resources.path("clinlp.resources", default_rules) as path:
-            self.add_rules(parse_rules(path))
+    @staticmethod
+    def _parse_qualifier(qualifier: str, qualifier_classes: dict[str, Qualifier]) -> Qualifier:
+        """
+        Parse a Qualifier from string.
+
+        Args:
+            qualifier: The qualifier (e.g. Negation.NEGATED).
+            qualifier_classes: A mapping of string to qualifier class.
+
+        Returns: A qualifier, as specified.
+        """
+
+        match_regexp = r"\w+\.\w+"
+
+        if not re.match(match_regexp, qualifier):
+            raise ValueError(
+                f"Cannot parse qualifier {qualifier}, please adhere to format "
+                f"{match_regexp} (e.g. NegationQualifier.NEGATED)"
+            )
+
+        qualifier_class, qualifier = qualifier.split(".")
+
+        return qualifier_classes[qualifier_class][qualifier]
+
+    @staticmethod
+    def _parse_direction(direction: str) -> ContextRuleDirection:
+        """
+        Parse a Context direction.
+
+        Args:
+            direction: The direction (e.g. preceding, following, etc.).
+
+        Returns: THe ContextRuleDirection.
+        """
+        return ContextRuleDirection[direction.upper()]
+
+    def _parse_rules(self, rules: Union[str | dict]) -> list[ContextRule]:
+        if isinstance(rules, str):
+            with open(rules, "rb") as file:
+                rules = json.load(file)
+
+        qualifiers = {
+            qualifier["qualifier"]: Qualifier(qualifier["qualifier"], qualifier["levels"])
+            for qualifier in rules["qualifiers"]
+        }
+
+        qualifier_rules = []
+
+        for rule in rules["rules"]:
+            qualifier = self._parse_qualifier(rule["qualifier"], qualifiers)
+            direction = self._parse_direction(rule["direction"])
+            max_scope = rule.get("max_scope", None)
+
+            qualifier_rules += [ContextRule(pattern, qualifier, direction, max_scope) for pattern in rule["patterns"]]
+
+        return qualifier_rules
 
     def add_rule(self, rule: ContextRule):
         """
@@ -290,7 +288,7 @@ class ContextMatcher(QualifierDetector):
 
     def __call__(self, doc: Doc):
         """
-        Apply the context matcher to a doc.
+        Apply the Context Algorithm to a doc.
         """
 
         if len(doc.ents) == 0:
@@ -318,7 +316,7 @@ class ContextMatcher(QualifierDetector):
                     rule=self._get_rule_from_match_id(match_id), start=start, end=end, offset=offset
                 )
 
-                pattern.set_initial_scope(sentence)
+                pattern.initialize_scope(sentence)
                 matched_patterns.append(pattern)
 
             match_scopes = self._compute_match_scopes(matched_patterns)
